refactor(jwt): split provider-proxy.service into per-class files
Extract RotatingShellSession and its supporting ws-rotation primitives
(RotationTracker, mergeSignals, isTokenExpiredMessage) into sibling
files. provider-proxy.service.ts re-exports ReceivedShellMessage so
existing consumer import paths are unchanged.
